ho creato il menu principale e dell'cliente modificato la struttura dei package come definito secondo https://www.html.it/pag/15098/cosa-sono-i-package-di-java/ e ho aggiunto il file note.txt dove ho messo la traccia e informazioni che potrebbero essere utili
