+++ Autonoleggio/src/it/terzaFila/autonoleggio/Batman.java
@@ -1,4 +1,4 @@
-package autonoleggio;
+package it.terzaFila.autonoleggio;
 
 public class Batman extends Utente{
